requirements.txt
PyTorch
numpy
tqdm
opencv-python
pyttsx3
pyyaml
nltk
pandas